Issue #57: A multi-platform feature to embed node.js
Instead of having multiple platform-specific feature, this
now provide a single feature with platform-specific bundles
that will be installed depending on the environment.

The main tricks was simply to set up environments in
target-platform-configuration in root pom.xml.

This was inspired by looking at the multiplatform.Reactor
integration-test of Tycho.