Find a way to configure Tern Modules that are default for Nature Adapters defined on a project
A new extension point elements defined for ternAdaptToNature element in ternNatureAdapters.exsd schema - 'defaultModules' and, its child, 'module' elements.
These elements allows to specify a list of Tern Modules (Libs and Plugins) that are to be loaded by default in Tern Project
in order to provide content assist and validation without a need of manual project configuration.

Signed-off-by: vrubezhny <vrubezhny@exadel.com>