Resolves #183 by resetting to original license
To satisfy the Apache License that covers the files from commons-io,
the original license headers have been added to the files, the full
text of the license has been included, and the NOTICE.txt from their
original distribution has been added, and a note appended to it
explaining which files it covers.

Disclaimer: I am not a lawyer, so I am not sure if there are any other
things that need to be done to satisfy the license. I am also not aware
of other files bundled here that should be covered by this license, or
by other licenses.